requirements.txt
torch~=2.7.1
torchvision~=0.22.1
opencv-python~=4.12.0.88
numpy~=2.2.6
flask~=3.1.1
mediapipe==0.10.14
joblib~=1.5.1
Pillow~=11.3.0
moviepy~=2.2.1
transformers~=4.55.0